+++ src/components/higlass/GenomicProfilesSubscriber.js
@@ -52,6 +52,7 @@ const REFERENCE_TILESETS = {
  * reference tilesets for the chromosome and gene annotations.
  * @param {string} props.assembly The genome assembly to use for the reference
  * tilesets for the chromosome and gene annotations.
+ * @param {string} props.title The title of the component.
  */
 export default function GenomicProfilesSubscriber(props) {
   const {
@@ -62,6 +63,7 @@ export default function GenomicProfilesSubscriber(props) {
     profileTrackNameKey = null,
     higlassServer = 'https://higlass.io/api/v1',
     assembly = 'hg38',
+    title = 'Genomic Profiles',
   } = props;
 
   // eslint-disable-next-line no-unused-vars
@@ -237,7 +239,7 @@ export default function GenomicProfilesSubscriber(props) {
   return (
     <div className="higlass-title-wrapper">
       <TitleInfo
-        title="Genomic Profiles"
+        title={title}
         removeGridComponent={removeGridComponent}
         theme={theme}
         isReady={isReady}